add rhea db switch

+++ libs/lib_sql_handler.py
@@ -2,14 +2,13 @@
 # Ultrabear 2020
 
 import sqlite3
-from sonnet_cfg import SQLITE3_LOCATION
 
 class db_error: # DB error codes
     OperationalError = sqlite3.OperationalError
 
 class db_handler:
-    def __init__(self):
-        self.con = sqlite3.connect(SQLITE3_LOCATION)
+    def __init__(self, db_location):
+        self.con = sqlite3.connect(db_location)
         self.cur = self.con.cursor()
 
     def __enter__(self):
@@ -48,6 +47,10 @@ class db_handler:
 
     def add_to_table(self, table, data):
 
+        # Test for attack
+        if table.count("\\") or table.count("'"):
+            raise db_error.OperationalError("Detected SQL injection attack")
+
         # Add insert data and generate base tables
         db_inputStr = f"REPLACE INTO '{table}' ("
         db_inputList = []
@@ -62,6 +65,10 @@ class db_handler:
 
     def fetch_rows_from_table(self, table, collumn_search):
 
+        # Test for attack
+        if table.count("\\") or table.count("'"):
+            raise db_error.OperationalError("Detected SQL injection attack")
+
         # Add SELECT data
         db_inputStr = f"SELECT * FROM '{table}' WHERE {collumn_search[0]} = ?"
         db_inputList = [collumn_search[1]]
@@ -73,6 +80,10 @@ class db_handler:
     # deletes rows from table where collumn i[0] has value i[1]
     def delete_rows_from_table(self, table, collumn_search): 
 
+        # Test for attack
+        if table.count("\\") or table.count("'"):
+            raise db_error.OperationalError("Detected SQL injection attack")
+
         # Do deletion setup
         db_inputStr = f"DELETE FROM '{table}' WHERE {collumn_search[0]}=?"
         db_inputList = [collumn_search[1]]
@@ -82,9 +93,17 @@ class db_handler:
 
     def delete_table(self, table): # drops the table specified
 
+        # Test for attack
+        if table.count("\\") or table.count("'"):
+            raise db_error.OperationalError("Detected SQL injection attack")
+
         self.cur.execute(f"DROP TABLE IF EXISTS '{table}';")
 
     def fetch_table(self, table): # Fetches a full table
+
+        # Test for attack
+        if table.count("\\") or table.count("'"):
+            raise db_error.OperationalError("Detected SQL injection attack")
 
         self.cur.execute(f"SELECT * FROM '{table}';")
 
@@ -102,95 +121,5 @@ class db_handler:
     def __exit__(self, err_type, err_value, err_traceback):
         self.con.commit()
         self.con.close()
-        if err_type:
-            raise err_type(err_value)
-
-
-# Because being lazy writes good code
-class db_hlapi:
-
-    def __init__(self, guild_id):
-        self.database = db_handler()
-        self.guild = guild_id
-
-    def __enter__(self):
-        return self
-
-    def grab_config(self, config):
-
-        try:
-            data = self.database.fetch_rows_from_table(f"{self.guild}_config", ["property",config])
-        except db_error.OperationalError:
-            data = []
-
-        if data:
-            return data[0][1]
-        else:
-            return []
-
-    def grab_user_infractions(self, userid):
-
-        try:
-            data = self.database.fetch_rows_from_table(f"{self.guild}_infractions", ["userID",userid])
-        except db_error.OperationalError:
-            data = []
-
-        return data
-
-    # Check if a message is on the starboard already
-    def in_starboard(self, message_id):
-        
-        try:
-            data = self.database.fetch_rows_from_table(f"{self.guild}_starboard", ["messageID", message_id])
-        except db_error.OperationalError:
-            data = True
-        
-        if data:
-            return True
-        else:
-            return False
-        
-    def add_to_starboard(self, message_id):
-
-        try:
-            self.database.add_to_table(f"{self.guild}_starboard", [["messageID", message_id]])
-        except db_error.OperationalError:
-            return False
-
-        return True
-
-    def grab_infraction(self, infractionID):
-
-        try:
-            infraction = self.database.fetch_rows_from_table(f"{self.guild}_infractions",["infractionID",infractionID])
-        except db_error.OperationalError:
-            infraction = None
-
-        if infraction:
-            return infraction[0]
-        else:
-            return False
-
-    def delete_infraction(self, infraction_id):
-
-        try:
-            self.database.delete_rows_from_table(f"{self.guild}_infractions",["infractionID",infraction_id])
-        except db_error.OperationalError:
-            pass
-
-    def mute_user(self, user, endtime, infractionID):
-
-        self.database.add_to_table(f"{self.guild}_mutes",[["infractionID", infractionID],["userID", user],["endMute",endtime]])
-
-    def unmute_user(self, infractionID):
-
-        self.database.delete_rows_from_table(f"{self.guild}_mutes", ["infractionID", infractionID])
-
-
-    def close(self):
-        self.database.close()
-
-    def __exit__(self, err_type, err_value, err_traceback):
-        self.database.close()
         if err_type:
             raise err_type(err_value)
